// frontend/src/api/model/getServiceOrdersIssueDescription.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * KS Tech Service Management API
 * API for managing users, Products, and service orders in a service-center / repair business. Designed using API-first principles.
 * OpenAPI spec version: 1.0.0
 */

export type GetServiceOrdersIssueDescription = typeof GetServiceOrdersIssueDescription[keyof typeof GetServiceOrdersIssueDescription];


export const GetServiceOrdersIssueDescription = {
  HARDWARE: 'HARDWARE',
  SOFTWARE: 'SOFTWARE',
  NETWORK: 'NETWORK',
  OTHER: 'OTHER',
} as const;